+++ src/test/groovy/ru/vyarus/java/generics/resolver/cases/cycle/CycledGeneric.java
@@ -1,0 +1,8 @@
+package ru.vyarus.java.generics.resolver.cases.cycle;
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 10.05.2019
+ */
+public interface CycledGeneric<T extends CycledGeneric<T>> {
+}